# docs/swagger.yaml
basePath: /
definitions:
  handler.HealthResponse:
    properties:
      providers:
        items:
          $ref: '#/definitions/handler.ProviderStatus'
        type: array
      status:
        type: string
      system:
        $ref: '#/definitions/handler.SystemInfo'
      timestamp:
        type: string
    type: object
  handler.ProviderStatus:
    properties:
      available:
        type: boolean
      name:
        type: string
    type: object
  handler.SystemInfo:
    properties:
      goroutines:
        type: integer
      memory_mb:
        type: number
      num_gc:
        type: integer
      uptime:
        type: string
    type: object
  model.AddressDetail:
    properties:
      building_name:
        description: 건물명
        type: string
      parcel_address:
        description: 지번 주소
        type: string
      road_address:
        description: 도로명 주소
        type: string
      zipcode:
        description: 우편번호
        type: string
    type: object
  model.BulkRequest:
    properties:
      addresses:
        description: 최대 100건
        items:
          type: string
        maxItems: 100
        type: array
    required:
    - addresses
    type: object
  model.BulkResponse:
    properties:
      processing_time_ms:
        type: integer
      results:
        items:
          $ref: '#/definitions/model.GeocodingResponse'
        type: array
      summary:
        properties:
          failed:
            type: integer
          success:
            type: integer
          total:
            type: integer
        type: object
    type: object
  model.Coordinate:
    properties:
      latitude:
        description: 위도 (y) - Decimal(9,6)
        type: number
      longitude:
        description: 경도 (x) - Decimal(9,6)
        type: number
    type: object
  model.GeocodingRequest:
    properties:
      address:
        description: 검색 주소
        type: string
      address_type:
        description: 주소 타입 (ROAD, PARCEL) - 선택적
        enum:
        - ROAD
        - PARCEL
        - road
        - parcel
        type: string
    required:
    - address
    type: object
  model.GeocodingResponse:
    properties:
      address_detail:
        $ref: '#/definitions/model.AddressDetail'
      attempts:
        description: Provider 시도 내역
        items:
          $ref: '#/definitions/model.ProviderAttempt'
        type: array
      coordinate:
        $ref: '#/definitions/model.Coordinate'
      error:
        type: string
      processed_at:
        type: string
      processing_time_ms:
        description: 밀리초
        type: integer
      provider:
        description: 최종 사용된 제공자
        type: string
      success:
        type: boolean
    type: object
  model.ProviderAttempt:
    properties:
      error:
        description: 에러 메시지
        type: string
      provider:
        description: Provider 이름
        type: string
      success:
        description: 성공 여부
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    한글 주소를 WGS84 좌표로 변환하는 하이브리드 지오코딩 API 서버입니다.
    vWorld와 Kakao API를 활용한 자동 폴백 시스템을 제공합니다.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: 하이브리드 지오코딩 API
  version: "1.0"
paths:
  /api/v1/geocode:
    post:
      consumes:
      - application/json
      description: |-
        한글 주소를 WGS84 좌표로 변환합니다. vWorld API를 우선 사용하고 실패 시 Kakao API로 자동 폴백됩니다.
        address_type을 지정하면 해당 타입(ROAD/PARCEL)으로만 검색합니다. 미지정 시 자동으로 ROAD → PARCEL 순서로 시도합니다.
      parameters:
      - description: '지오코딩 요청 (address_type은 선택사항: ROAD 또는 PARCEL)'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.GeocodingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 변환 성공
          schema:
            $ref: '#/definitions/model.GeocodingResponse'
        "400":
          description: 잘못된 요청
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 주소를 찾을 수 없음
          schema:
            $ref: '#/definitions/model.GeocodingResponse'
        "500":
          description: 서버 에러
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 주소를 좌표로 변환
      tags:
      - geocoding
  /api/v1/geocode/bulk:
    post:
      consumes:
      - application/json
      description: 여러 한글 주소를 WGS84 좌표로 변환합니다. 최대 100개까지 처리 가능하며, 최대 10개씩 동시 처리됩니다.
      parameters:
      - description: 대량 지오코딩 요청 (최대 100개)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.BulkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 변환 결과
          schema:
            $ref: '#/definitions/model.BulkResponse'
        "400":
          description: 잘못된 요청 (빈 배열 또는 100개 초과)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 서버 에러
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 여러 주소를 좌표로 변환
      tags:
      - geocoding
  /health:
    get:
      description: 서비스와 Provider들의 상태를 확인합니다. 시스템 정보(메모리, Goroutine 등)도 함께 제공됩니다.
      produces:
      - application/json
      responses:
        "200":
          description: 서비스 정상
          schema:
            $ref: '#/definitions/handler.HealthResponse'
        "503":
          description: 서비스 비정상 (Provider 장애)
          schema:
            $ref: '#/definitions/handler.HealthResponse'
      summary: 서비스 상태 확인
      tags:
      - health
  /ping:
    get:
      description: 서비스가 살아있는지 간단히 확인합니다
      produces:
      - application/json
      responses:
        "200":
          description: pong 응답
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Ping
      tags:
      - health
  /ready:
    get:
      description: 서비스가 요청을 처리할 준비가 되었는지 확인합니다. Kubernetes Readiness Probe에 사용할 수
        있습니다.
      produces:
      - application/json
      responses:
        "200":
          description: 준비 완료
          schema:
            additionalProperties:
              type: boolean
            type: object
        "503":
          description: 준비 안됨
          schema:
            additionalProperties:
              type: boolean
            type: object
      summary: Readiness 체크
      tags:
      - health
schemes:
- http
- https
swagger: "2.0"
tags:
- description: 지오코딩 API
  name: geocoding
- description: 헬스체크 API
  name: health